Выделил логику авторизации в отдельный сервис
